Working on Attacking
I am working on getting the AI to attack a ship if it can. I have set up two AI states, one ReadyToShoot and one Shooting. The ReadyToShoot function turns to shoot. It angles itself between 80 and 100 relative to the enemy. Then it shoots. I am trying to get it so that it knows what side it needs to shoot from after it turns and then checks to see if it can shoot again. I am thinking it needs to do ReadyToShoot in between every shot, and that can be how it checks if it can shoot again.